fiz(bias): alter to float field

+++ crowd/crowd/tasks/models.py
@@ -21,7 +21,7 @@ class TaskAssignment(TimeStampedModel):
 	task = models.ForeignKey(Task, related_name='assignments')
 	user_answer = models.ForeignKey('AnswerOption', null=True)
 	user_answer_time = models.DateTimeField(null=True)
-	bias_at_answer = models.IntegerField(_("bias"), default=0)
+	bias_at_answer = models.FloatField(_("bias"), default=0)
 
 	maximum = 10
 	middle = 0
